@@ -6,6 +6,7 @@ import 'package:receptico/generated/l10n.dart';
 import 'package:receptico/core/UI/theme.dart';
 
 import '../bloc/bloc.dart';
+import '../common/enum.dart';
 import '../widget/widget.dart';
 
 @RoutePage()
@@ -17,103 +18,158 @@ class RestorePhonePage extends StatefulWidget {
 }
 
 class _RestorePhonePageState extends State<RestorePhonePage> {
-  late bool _isError;
-  AuthLoginFailState _authState = AuthLoginFailState();
-  final _emailOrPhoneController = TextEditingController();
+  final Map<EInput, TextEditingController> _controllers = {
+    EInput.smsCode: TextEditingController(),
+    EInput.password: TextEditingController(),
+    EInput.confirmPassword: TextEditingController(),
+  };
+
+  void _clearForm() =>
+      _controllers.forEach((key, controller) => controller.clear());
 
   @override
-  void initState() {
-    super.initState();
-    _isError = false;
+  void dispose() {
+    _controllers.forEach((key, controller) => controller.dispose());
+    super.dispose();
   }
 
-  void _clearFailMessage() {
-    if (_isError) {
-      _isError = false;
-      setState(() => _authState = AuthLoginFailState());
-    }
+  String? _smsCodeValidate(String? value) {
+    context.read<AuthBloc>().add(
+          AuthSMSCodeValidateEvent(
+            value: value,
+            localization: S.of(context),
+          ),
+        );
+    return value;
   }
 
-  void _submit() {}
+  String? _passwordValidate(String? value) {
+    _confirmPasswordValidate(_controllers[EInput.confirmPassword]?.text);
+    context.read<AuthBloc>().add(
+          AuthPasswordValidateEvent(
+            value: value,
+            localization: S.of(context),
+          ),
+        );
+    return value;
+  }
+
+  String? _confirmPasswordValidate(String? value) {
+    context.read<AuthBloc>().add(
+          AuthConfirmPasswordValidateEvent(
+            value: _controllers[EInput.password]?.text,
+            confirmValue: value,
+            localization: S.of(context),
+          ),
+        );
+    return value;
+  }
+
+  void _submit() {
+    _smsCodeValidate(_controllers[EInput.smsCode]?.text);
+    _passwordValidate(_controllers[EInput.password]?.text);
+    _confirmPasswordValidate(_controllers[EInput.confirmPassword]?.text);
+
+    context.read<AuthBloc>().add(
+          AuthRestorePhoneEvent(
+            code: _controllers[EInput.smsCode]?.text ?? '',
+            password: _controllers[EInput.password]?.text ?? '',
+            confirmPassword: _controllers[EInput.confirmPassword]?.text ?? '',
+            localization: S.of(context),
+          ),
+        );
+  }
 
   @override
   Widget build(BuildContext context) {
     final router = AutoRouter.of(context);
     final localization = S.of(context);
-    return Scaffold(
-      body: BlocListener<AuthBloc, AuthState>(
-        listener: (context, state) {
-          if (state is AuthLoginFailState) {
-            _isError = true;
-            setState(() => _authState = state);
-          }
-        },
-        child: Stack(
-          children: [
-            const ScreenBackgroundWidget(),
-            Container(
-              alignment: Alignment.center,
-              child: SizedBox(
-                width: 350,
-                height: 500,
-                child: SingleChildScrollView(
-                  child: Column(
-                    children: [
-                      TitleWidget(
-                        title: localization.restorePasswordTitle,
-                        subtitle: localization.smsSubtitle,
-                      ),
-                      const SizedBox(height: 37),
-                      SizedBox(
-                        width: 278,
-                        child: Column(
-                          children: [
-                            PhoneWidget(text: '+380********98'),
-                            TextInputWidget(
-                              controller: _emailOrPhoneController,
-                              placeholder: localization.smsCodePlaceholder,
-                              onTap: _clearFailMessage,
-                              errorMessage: _authState.emailOrPhoneError,
-                              margin: EdgeInsets.only(bottom: 16.0),
-                              marginWithError: EdgeInsets.only(bottom: 10.0),
-                            ),
-                            TextInputWidget(
-                              controller: _emailOrPhoneController,
-                              placeholder: localization.newPasswordPlaceholder,
-                              onTap: _clearFailMessage,
-                              errorMessage: _authState.emailOrPhoneError,
-                              margin: EdgeInsets.only(bottom: 16.0),
-                              marginWithError: EdgeInsets.only(bottom: 10.0),
-                            ),
-                            TextInputWidget(
-                              controller: _emailOrPhoneController,
-                              placeholder:
-                                  localization.confirmPasswordPlaceholder,
-                              onTap: _clearFailMessage,
-                              errorMessage: _authState.emailOrPhoneError,
-                              margin: EdgeInsets.only(bottom: 25.0),
-                              marginWithError: EdgeInsets.only(bottom: 22.0),
-                            ),
-                            TextButtonWidget(
-                              height: 50,
-                              text: localization.confirmPasswordChangeButton,
-                              onPressed: () => router.navigate(
-                                  const RestorePhoneRoute()), //_submit,
-                            ),
-                            const SizedBox(height: 28),
-                            BackLinkWidget(
-                              text: localization.returnToLoginLink,
-                              onTap: () => router.navigate(const LoginRoute()),
-                            ),
-                          ],
+    return WillPopScope(
+      onWillPop: () async {
+        context.read<AuthBloc>().add(AuthRouteEvent());
+        return true;
+      },
+      child: Scaffold(
+        body: BlocListener<AuthBloc, AuthState>(
+          listener: (context, state) {
+            // if (state is AuthLoginFailState) {
+            //   //_isError = true;
+            //   //setState(() => _authState = state);
+            // }
+          },
+          child: Stack(
+            children: [
+              const ScreenBackgroundWidget(),
+              Container(
+                alignment: Alignment.center,
+                child: SizedBox(
+                  width: 350,
+                  height: 500,
+                  child: SingleChildScrollView(
+                    child: Column(
+                      children: [
+                        TitleWidget(
+                          title: localization.restorePasswordTitle,
+                          subtitle: localization.smsSubtitle,
                         ),
-                      ),
-                    ],
+                        const SizedBox(height: 37),
+                        SizedBox(
+                          width: 278,
+                          child: Column(
+                            children: [
+                              PhoneWidget(text: '+380********98'),
+                              SelectorTextInputWidget(
+                                placeholder: localization.smsCodePlaceholder,
+                                controller: _controllers[EInput.smsCode],
+                                selector: (state) =>
+                                    state.errors?[EBlocError.smsCode],
+                                onChanged: _smsCodeValidate,
+                              ),
+                              SelectorPasswordInputWidget(
+                                placeholder: localization.passwordPlaceholder,
+                                controller: _controllers[EInput.password],
+                                selector: (state) =>
+                                    state.errors?[EBlocError.password],
+                                onChanged: _passwordValidate,
+                                margin: EdgeInsets.only(bottom: 16.0),
+                                marginWithError: EdgeInsets.only(bottom: 10.0),
+                              ),
+                              SelectorPasswordInputWidget(
+                                placeholder:
+                                    localization.confirmPasswordPlaceholder,
+                                controller:
+                                    _controllers[EInput.confirmPassword],
+                                selector: (state) =>
+                                    state.errors?[EBlocError.confirmPassword],
+                                onChanged: _confirmPasswordValidate,
+                                margin: EdgeInsets.only(bottom: 28.0),
+                              ),
+                              TextButtonWidget(
+                                height: 50,
+                                text: localization.confirmPasswordChangeButton,
+                                onPressed: _submit,
+                              ),
+                              const SizedBox(height: 28),
+                              BackLinkWidget(
+                                text: localization.returnToLoginLink,
+                                onTap: () {
+                                  router.navigate(const LoginRoute());
+                                  context
+                                      .read<AuthBloc>()
+                                      .add(AuthRouteEvent());
+                                },
+                              ),
+                            ],
+                          ),
+                        ),
+                      ],
+                    ),
                   ),
                 ),
               ),
-            ),
-          ],
+              LoadingWidget(),
+            ],
+          ),
         ),
       ),
     );
@@ -141,7 +197,7 @@ class PhoneWidget extends StatelessWidget {
       margin: margin,
       padding: padding,
       decoration: BoxDecoration(
-        color: context.color.background.dashboarPhone,
+        color: context.color.background.dashboarPhone.safe,
         border: Border.all(
           color: context.color.border.dashboarPhone.safe,
           width: 1.0,
@@ -151,7 +207,7 @@ class PhoneWidget extends StatelessWidget {
       child: Text(
         text,
         style: context.font.body?.copyWith(
-          color: context.color.font.input,
+          color: context.color.font.input.safe,
         ),
       ),
     );
