@@ -3,4 +3,5 @@ export 'login_page.dart';
 export 'register_page.dart';
 export 'restore_password_page.dart';
 export 'restore_phone_page.dart';
-export 'restore_email_page.dart';
+export 'send_email_page.dart';
+export 'fail_page.dart';
