@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:receptico/core/UI/color/color.dart';
 
 import 'text_input_widget.dart';
 
@@ -7,11 +8,11 @@ class PasswordInputWidget extends StatefulWidget {
   final double width, height, borderRadius, fontSize;
   final EdgeInsets margin, marginWithError;
   final String placeholderText;
-  final String? errorMessage;
-  final Color placeholderColor, inputTextColor, backgroundColor;
+  final String? Function(String?)? onChanged;
   final VoidCallback? onTap;
   final TextEditingController? controller;
   final Widget? helpWidget;
+  final String? error;
 
   const PasswordInputWidget({
     super.key,
@@ -22,13 +23,11 @@ class PasswordInputWidget extends StatefulWidget {
     this.fontSize = 17,
     this.marginWithError = const EdgeInsets.all(0),
     this.margin = const EdgeInsets.all(0),
-    this.placeholderColor = const Color(0xFF8E8E93),
-    this.inputTextColor = Colors.black,
-    this.backgroundColor = Colors.white,
     this.onTap,
     this.controller,
-    this.errorMessage,
     this.helpWidget,
+    this.onChanged,
+    this.error,
   });
 
   @override
@@ -40,40 +39,34 @@ class _PasswordInputWidgetState extends State<PasswordInputWidget> {
 
   @override
   Widget build(BuildContext context) {
-    const offset = Offset(0, 0);
-    const padding = EdgeInsets.only(left: 16, right: 44, top: 10, bottom: 10),
-        iconPadding = EdgeInsets.only(bottom: 9);
-    return Container(
-      margin:
-          widget.errorMessage == null ? widget.margin : widget.marginWithError,
-      child: Stack(
-        children: [
-          TextInputWidget(
-            controller: widget.controller,
-            placeholder: widget.placeholderText,
-            obscureText: _isObscured,
-            padding: padding,
-            errorMessage: widget.errorMessage,
-            onTap: widget.onTap,
-            helpWidget: widget.helpWidget,
-          ),
-          Container(
-            alignment: Alignment.centerRight,
-            child: Transform.translate(
-              offset: offset,
-              child: IconButton(
-                padding: iconPadding,
-                onPressed: () => setState(() => _isObscured = !_isObscured),
-                icon: Icon(
-                  _isObscured ? CupertinoIcons.eye_slash : CupertinoIcons.eye,
-                  size: widget.fontSize,
-                  color: widget.inputTextColor,
-                ),
-              ),
+    const padding = EdgeInsets.only(left: 16, right: 44, top: 10, bottom: 10);
+    return Stack(
+      children: [
+        TextInputWidget(
+          onChanged: widget.onChanged,
+          controller: widget.controller,
+          placeholder: widget.placeholderText,
+          obscureText: _isObscured,
+          padding: padding,
+          onTap: widget.onTap,
+          helpWidget: widget.helpWidget,
+          error: widget.error,
+          margin: widget.margin,
+          marginWithError: widget.marginWithError,
+        ),
+        Container(
+          height: widget.height,
+          alignment: Alignment.centerRight,
+          child: IconButton(
+            onPressed: () => setState(() => _isObscured = !_isObscured),
+            icon: Icon(
+              _isObscured ? CupertinoIcons.eye_slash : CupertinoIcons.eye,
+              size: widget.fontSize,
+              color: context.color.font.input,
             ),
           ),
-        ],
-      ),
+        ),
+      ],
     );
   }
 }
