@@ -21,40 +21,72 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
   static String m0(email) =>
+      "If the account “${email}” exists, instructions for verifying the email will be sent to your email.";
+
+  static String m1(min) => "Password must be at least ${min} characters long";
+
+  static String m2(min) => "Username must be at least ${min} characters long";
+
+  static String m3(email) =>
       "If the account ${email} exists, password recovery instructions will be sent to your email.";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "accountQuestion":
             MessageLookupByLibrary.simpleMessage("Already have an account?"),
+        "confirmEmailMessage": m0,
         "confirmPasswordChangeButton":
             MessageLookupByLibrary.simpleMessage("Change password "),
         "confirmPasswordPlaceholder":
             MessageLookupByLibrary.simpleMessage("Confirm password"),
         "createAccount": MessageLookupByLibrary.simpleMessage("Create now"),
+        "emailExistError":
+            MessageLookupByLibrary.simpleMessage("Email already in use"),
         "emailOrPhonePlaceholder":
             MessageLookupByLibrary.simpleMessage("Email/Phone"),
         "forgottenPassword":
             MessageLookupByLibrary.simpleMessage("Forgot password?"),
+        "invalidCharacterEnteredError":
+            MessageLookupByLibrary.simpleMessage("Invalid character entered"),
+        "invalidEmailError":
+            MessageLookupByLibrary.simpleMessage("Invalid email"),
+        "invalidPhoneError":
+            MessageLookupByLibrary.simpleMessage("Invalid phone number"),
+        "invalidVerificationCodeError":
+            MessageLookupByLibrary.simpleMessage("Invalid verification code"),
         "loginButton": MessageLookupByLibrary.simpleMessage("Sign in"),
         "loginLink": MessageLookupByLibrary.simpleMessage("Sign in"),
         "loginSubtitle":
             MessageLookupByLibrary.simpleMessage("Log in to your account"),
         "loginTitle": MessageLookupByLibrary.simpleMessage("Sign in"),
+        "minPasswordLengError": m1,
+        "minUsernameLengError": m2,
         "namePlaceholder": MessageLookupByLibrary.simpleMessage("Name"),
         "newPasswordPlaceholder":
             MessageLookupByLibrary.simpleMessage("New password"),
         "noAccountQuestion":
             MessageLookupByLibrary.simpleMessage("Don\'t have an account?"),
+        "oopsTitle":
+            MessageLookupByLibrary.simpleMessage("Oops, something went wrong"),
+        "passwordLowercaseError": MessageLookupByLibrary.simpleMessage(
+            "Password must contain at least one lowercase letter"),
+        "passwordNumberError": MessageLookupByLibrary.simpleMessage(
+            "Password must contain at least one number"),
         "passwordPlaceholder": MessageLookupByLibrary.simpleMessage("Password"),
+        "passwordSpecialCharacterError": MessageLookupByLibrary.simpleMessage(
+            "Password must contain at least one special character (_-./\\)"),
+        "passwordUppercaseError": MessageLookupByLibrary.simpleMessage(
+            "Password must contain at least one uppercase letter"),
+        "passwordsNoMatchError":
+            MessageLookupByLibrary.simpleMessage("Passwords do not match"),
         "registerButton": MessageLookupByLibrary.simpleMessage("Sign up"),
         "registerLink": MessageLookupByLibrary.simpleMessage("Sign up"),
         "registerSubtitle":
             MessageLookupByLibrary.simpleMessage("Create a new account"),
         "registerTitle": MessageLookupByLibrary.simpleMessage("Sign up"),
-        "requiredEmailError":
-            MessageLookupByLibrary.simpleMessage("Введіть пошту"),
-        "restoreEmailMessage": m0,
+        "requiredError": MessageLookupByLibrary.simpleMessage(
+            "The field must not be empty."),
+        "restoreEmailMessage": m3,
         "restorePasswordButton":
             MessageLookupByLibrary.simpleMessage("Restore password"),
         "restorePasswordSubtitle": MessageLookupByLibrary.simpleMessage(
@@ -67,7 +99,16 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("Confirmation code"),
         "smsSubtitle": MessageLookupByLibrary.simpleMessage(
             "An SMS message with a confirmation code to recover your password has been sent to your phone number."),
+        "tryAgainButton": MessageLookupByLibrary.simpleMessage("Try again"),
+        "userDisabledError":
+            MessageLookupByLibrary.simpleMessage("User disabled"),
+        "userNoVerifiedError":
+            MessageLookupByLibrary.simpleMessage("User not verified"),
+        "userNotFoundError":
+            MessageLookupByLibrary.simpleMessage("User not found"),
         "wantToCookQuestion": MessageLookupByLibrary.simpleMessage(
-            "Do you want to cook something?")
+            "Do you want to cook something?"),
+        "wrongPasswordError":
+            MessageLookupByLibrary.simpleMessage("Wrong password")
       };
 }
