@@ -21,40 +21,72 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'uk';
 
   static String m0(email) =>
+      "Якщо обліковий запис “${email}” існує, на електронну пошту буде надіслано інструкцію з підтвердження пошти.";
+
+  static String m1(min) => "Пароль має бути не менше ${min} символів";
+
+  static String m2(min) => "Ім’я не може бути меньшим ніж ${min} символів";
+
+  static String m3(email) =>
       "Якщо обліковий запис “${email}” існує, на електронну пошту буде надіслано інструкцію з відновлення паролю.";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
         "accountQuestion":
             MessageLookupByLibrary.simpleMessage("Вже маєте акаунт?"),
+        "confirmEmailMessage": m0,
         "confirmPasswordChangeButton":
             MessageLookupByLibrary.simpleMessage("Підвердити зміну пароля"),
         "confirmPasswordPlaceholder":
             MessageLookupByLibrary.simpleMessage("Підтвердження паролю"),
         "createAccount": MessageLookupByLibrary.simpleMessage("Створити зараз"),
+        "emailExistError": MessageLookupByLibrary.simpleMessage(
+            "Цей електронний лист вже використовується"),
         "emailOrPhonePlaceholder":
             MessageLookupByLibrary.simpleMessage("Email/Телефон"),
         "forgottenPassword":
             MessageLookupByLibrary.simpleMessage("Забули пароль?"),
+        "invalidCharacterEnteredError":
+            MessageLookupByLibrary.simpleMessage("Введено недійсний символ"),
+        "invalidEmailError":
+            MessageLookupByLibrary.simpleMessage("Недійсна пошта"),
+        "invalidPhoneError":
+            MessageLookupByLibrary.simpleMessage("Недійсний номер телефону"),
+        "invalidVerificationCodeError":
+            MessageLookupByLibrary.simpleMessage("Невірний код підтвердження"),
         "loginButton": MessageLookupByLibrary.simpleMessage("Увійти"),
         "loginLink": MessageLookupByLibrary.simpleMessage("Увійти"),
         "loginSubtitle": MessageLookupByLibrary.simpleMessage(
             "Увійдіть до облікового запису"),
         "loginTitle": MessageLookupByLibrary.simpleMessage("Вхід"),
+        "minPasswordLengError": m1,
+        "minUsernameLengError": m2,
         "namePlaceholder": MessageLookupByLibrary.simpleMessage("Ім’я"),
         "newPasswordPlaceholder":
             MessageLookupByLibrary.simpleMessage("Новий пароль"),
         "noAccountQuestion":
             MessageLookupByLibrary.simpleMessage("Не маєте акаунту?"),
+        "oopsTitle":
+            MessageLookupByLibrary.simpleMessage("Ой, щось пішло не так"),
+        "passwordLowercaseError": MessageLookupByLibrary.simpleMessage(
+            "Пароль має містити принаймні одну малу літеру"),
+        "passwordNumberError": MessageLookupByLibrary.simpleMessage(
+            "Пароль повинен містити хоча б одне число"),
         "passwordPlaceholder": MessageLookupByLibrary.simpleMessage("Пароль"),
+        "passwordSpecialCharacterError": MessageLookupByLibrary.simpleMessage(
+            "Пароль повинен містити хоча б один спеціальний символ (_-./\\)"),
+        "passwordUppercaseError": MessageLookupByLibrary.simpleMessage(
+            "Пароль має містити принаймні одну велику літеру"),
+        "passwordsNoMatchError":
+            MessageLookupByLibrary.simpleMessage("Паролі не співпадають"),
         "registerButton": MessageLookupByLibrary.simpleMessage("Створити"),
         "registerLink": MessageLookupByLibrary.simpleMessage("Зареєструватися"),
         "registerSubtitle":
             MessageLookupByLibrary.simpleMessage("Створіть новий акаунт"),
         "registerTitle": MessageLookupByLibrary.simpleMessage("Реєстрація"),
-        "requiredEmailError":
-            MessageLookupByLibrary.simpleMessage("Введіть пошту"),
-        "restoreEmailMessage": m0,
+        "requiredError":
+            MessageLookupByLibrary.simpleMessage("Поле не має бути пустим"),
+        "restoreEmailMessage": m3,
         "restorePasswordButton":
             MessageLookupByLibrary.simpleMessage("Відновлення паролю"),
         "restorePasswordSubtitle": MessageLookupByLibrary.simpleMessage(
@@ -67,7 +99,17 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("Код підтвердження"),
         "smsSubtitle": MessageLookupByLibrary.simpleMessage(
             "На Ваш номер телефону відправлено SMS-повідомлення з кодом підтвердження для відновлення пароля"),
+        "tryAgainButton":
+            MessageLookupByLibrary.simpleMessage("Спробувати знову"),
+        "userDisabledError":
+            MessageLookupByLibrary.simpleMessage("Користувач заблокований"),
+        "userNoVerifiedError":
+            MessageLookupByLibrary.simpleMessage("Користувача не підтверджено"),
+        "userNotFoundError":
+            MessageLookupByLibrary.simpleMessage("Користувача не знайдено"),
         "wantToCookQuestion":
-            MessageLookupByLibrary.simpleMessage("Хочеш Щось Приготувати?")
+            MessageLookupByLibrary.simpleMessage("Хочеш Щось Приготувати?"),
+        "wrongPasswordError":
+            MessageLookupByLibrary.simpleMessage("Невірний пароль")
       };
 }
